refactor: move validation audit out of webhooks package
