+++ pkg/webhooks/resource/audit/validate_audit.go
@@ -1,4 +1,4 @@
-package resource
+package audit
 
 import (
 	"strings"
