+++ pkg/webhooks/resource/handlers.go
@@ -27,6 +27,7 @@ import (
 	engineutils "github.com/kyverno/kyverno/pkg/utils/engine"
 	jsonutils "github.com/kyverno/kyverno/pkg/utils/json"
 	"github.com/kyverno/kyverno/pkg/webhooks"
+	"github.com/kyverno/kyverno/pkg/webhooks/resource/audit"
 	"github.com/kyverno/kyverno/pkg/webhooks/resource/validation"
 	webhookgenerate "github.com/kyverno/kyverno/pkg/webhooks/updaterequest"
 	webhookutils "github.com/kyverno/kyverno/pkg/webhooks/utils"
@@ -59,7 +60,7 @@ type handlers struct {
 	prGenerator       policyreport.GeneratorInterface
 	urGenerator       webhookgenerate.Generator
 	eventGen          event.Interface
-	auditHandler      AuditHandler
+	auditHandler      audit.AuditHandler
 	openAPIController openapi.ValidateInterface
 	pcBuilder         webhookutils.PolicyContextBuilder
 	urUpdater         webhookutils.UpdateRequestUpdater
@@ -78,7 +79,7 @@ func NewHandlers(
 	prGenerator policyreport.GeneratorInterface,
 	urGenerator webhookgenerate.Generator,
 	eventGen event.Interface,
-	auditHandler AuditHandler,
+	auditHandler audit.AuditHandler,
 	openAPIController openapi.ValidateInterface,
 ) webhooks.Handlers {
 	return &handlers{
